Fixes to allow enable osdf defaults
This modifies the plugin to allow enabling osdf defaults for
stashcp/stash_plugin. The issue was that we were always using the
default osdf discovery url so the test ignored `pelican.example.com` and
just went ahead and downloaded the testfile. This is fixed now and we
will fail if the discovery url is not an actual url